[API] Change endpoint for requesting customer authentication token
