+++ src/Sylius/Bundle/ApiBundle/Resources/config/integrations/swagger.xml
@@ -74,12 +74,8 @@
             <tag name="sylius.open_api.modifier" />
         </service>
 
-        <service id="Sylius\Bundle\ApiBundle\OpenApi\Documentation\ShopAuthenticationTokenDocumentationModifier">
-            <argument>%sylius.security.new_api_route%</argument>
-            <tag name="sylius.open_api.modifier" />
-        </service>
-
         <service id="Sylius\Bundle\ApiBundle\OpenApi\Documentation\CustomerDocumentationModifier">
+            <argument>%sylius.security.new_api_route%</argument>
             <tag name="sylius.open_api.modifier" />
         </service>
